fix: handle when final step is missing newline
Prior to this change, if the final line of the recipe didn't contain a
new line character, the parser would hang. This would occur when there
are no more TAGs to parse and it's just straight text between the
buffer head and the end of the buffer. If there were no more tags the
rest of the text was returned, however the buffer position wasn't
correctly updated to point to the end of the buffer.

This change will update the buffers position to point to the end of the
string when there are no more tags left.